# open-api/immich-openapi-specs.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/sync/acknowledge": {
      "post": {
        "operationId": "sendSyncAck",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "type": "string"
                },
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Sync"
        ]
      }
    },
    "/sync/delta-sync": {
      "post": {
        "operationId": "getDeltaSync",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetDeltaSyncDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetDeltaSyncResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Sync"
        ]
      }
    },
    "/sync/full-sync": {
      "post": {
        "operationId": "getFullSyncForUser",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetFullSyncDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AssetResponseDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Sync"
        ]
      }
    },
    "/sync/stream": {
      "post": {
        "operationId": "getSyncStream",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncStreamDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Sync"
        ]
      }
    },
...
  },
  "info": {
    "title": "Immich",
    "description": "Immich API",
    "version": "1.126.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "Bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "immich_access_token"
      },
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
...
      "AssetDeltaSyncDto": {
        "properties": {
          "updatedAfter": {
            "format": "date-time",
            "type": "string"
          },
          "userIds": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "updatedAfter",
          "userIds"
        ],
        "type": "object"
      },
      "AssetDeltaSyncResponseDto": {
        "properties": {
          "deleted": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "needsFullSync": {
            "type": "boolean"
          },
          "upserted": {
            "items": {
              "$ref": "#/components/schemas/AssetResponseDto"
            },
            "type": "array"
          }
        },
        "required": [
          "deleted",
          "needsFullSync",
          "upserted"
        ],
        "type": "object"
      },
...
      "AssetFaceWithoutPersonResponseDto": {
        "properties": {
          "boundingBoxX1": {
            "type": "integer"
          },
          "boundingBoxX2": {
            "type": "integer"
          },
          "boundingBoxY1": {
            "type": "integer"
          },
          "boundingBoxY2": {
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "imageHeight": {
            "type": "integer"
          },
          "imageWidth": {
            "type": "integer"
          },
          "sourceType": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SourceType"
              }
            ]
          }
        },
        "required": [
          "boundingBoxX1",
          "boundingBoxX2",
          "boundingBoxY1",
          "boundingBoxY2",
          "id",
          "imageHeight",
          "imageWidth"
        ],
        "type": "object"
      },
      "AssetFullSyncDto": {
        "properties": {
          "lastId": {
            "format": "uuid",
            "type": "string"
          },
          "limit": {
            "minimum": 1,
            "type": "integer"
          },
          "updatedUntil": {
            "format": "date-time",
            "type": "string"
          },
          "userId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "limit",
          "updatedUntil"
        ],
        "type": "object"
      },
...
      "AssetResponseDto": {
        "properties": {
          "checksum": {
            "description": "base64 encoded sha1 hash",
            "type": "string"
          },
          "deviceAssetId": {
            "type": "string"
          },
          "deviceId": {
            "type": "string"
          },
          "duplicateId": {
            "nullable": true,
            "type": "string"
          },
          "duration": {
            "type": "string"
          },
          "exifInfo": {
            "$ref": "#/components/schemas/ExifResponseDto"
          },
          "fileCreatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "fileModifiedAt": {
            "format": "date-time",
            "type": "string"
          },
          "hasMetadata": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "isArchived": {
            "type": "boolean"
          },
          "isFavorite": {
            "type": "boolean"
          },
          "isOffline": {
            "type": "boolean"
          },
          "isTrashed": {
            "type": "boolean"
          },
          "libraryId": {
            "deprecated": true,
            "description": "This property was deprecated in v1.106.0",
            "nullable": true,
            "type": "string"
          },
          "livePhotoVideoId": {
            "nullable": true,
            "type": "string"
          },
          "localDateTime": {
            "format": "date-time",
            "type": "string"
          },
          "originalFileName": {
            "type": "string"
          },
          "originalMimeType": {
            "type": "string"
          },
          "originalPath": {
            "type": "string"
          },
          "owner": {
            "$ref": "#/components/schemas/UserResponseDto"
          },
          "ownerId": {
            "type": "string"
          },
          "people": {
            "items": {
              "$ref": "#/components/schemas/PersonWithFacesResponseDto"
            },
            "type": "array"
          },
          "resized": {
            "deprecated": true,
            "description": "This property was deprecated in v1.113.0",
            "type": "boolean"
          },
          "stack": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AssetStackResponseDto"
              }
            ],
            "nullable": true
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/TagResponseDto"
            },
            "type": "array"
          },
          "thumbhash": {
            "nullable": true,
            "type": "string"
          },
          "type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AssetTypeEnum"
              }
            ]
          },
          "unassignedFaces": {
            "items": {
              "$ref": "#/components/schemas/AssetFaceWithoutPersonResponseDto"
            },
            "type": "array"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "checksum",
          "deviceAssetId",
          "deviceId",
          "duration",
          "fileCreatedAt",
          "fileModifiedAt",
          "hasMetadata",
          "id",
          "isArchived",
          "isFavorite",
          "isOffline",
          "isTrashed",
          "localDateTime",
          "originalFileName",
          "originalPath",
          "ownerId",
          "thumbhash",
          "type",
          "updatedAt"
        ],
        "type": "object"
      },
      "AssetStackResponseDto": {
        "properties": {
          "assetCount": {
            "type": "integer"
          },
          "id": {
            "type": "string"
          },
          "primaryAssetId": {
            "type": "string"
          }
        },
        "required": [
          "assetCount",
          "id",
          "primaryAssetId"
        ],
        "type": "object"
      },
...
      "AssetTypeEnum": {
        "enum": [
          "IMAGE",
          "VIDEO",
          "AUDIO",
          "OTHER"
        ],
        "type": "string"
      },
...
      "ExifResponseDto": {
        "properties": {
          "city": {
            "default": null,
            "nullable": true,
            "type": "string"
          },
          "country": {
            "default": null,
            "nullable": true,
            "type": "string"
          },
          "dateTimeOriginal": {
            "default": null,
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "description": {
            "default": null,
            "nullable": true,
            "type": "string"
          },
          "exifImageHeight": {
            "default": null,
            "nullable": true,
            "type": "number"
          },
          "exifImageWidth": {
            "default": null,
            "nullable": true,
            "type": "number"
          },
          "exposureTime": {
            "default": null,
            "nullable": true,
            "type": "string"
          },
          "fNumber": {
            "default": null,
            "nullable": true,
            "type": "number"
          },
          "fileSizeInByte": {
            "default": null,
            "format": "int64",
            "nullable": true,
            "type": "integer"
          },
          "focalLength": {
            "default": null,
            "nullable": true,
            "type": "number"
          },
          "iso": {
            "default": null,
            "nullable": true,
            "type": "number"
          },
          "latitude": {
            "default": null,
            "nullable": true,
            "type": "number"
          },
          "lensModel": {
            "default": null,
            "nullable": true,
            "type": "string"
          },
          "longitude": {
            "default": null,
            "nullable": true,
            "type": "number"
          },
          "make": {
            "default": null,
            "nullable": true,
            "type": "string"
          },
          "model": {
            "default": null,
            "nullable": true,
            "type": "string"
          },
          "modifyDate": {
            "default": null,
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "orientation": {
            "default": null,
            "nullable": true,
            "type": "string"
          },
          "projectionType": {
            "default": null,
            "nullable": true,
            "type": "string"
          },
          "rating": {
            "default": null,
            "nullable": true,
            "type": "number"
          },
          "state": {
            "default": null,
            "nullable": true,
            "type": "string"
          },
          "timeZone": {
            "default": null,
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "PersonWithFacesResponseDto": {
        "properties": {
          "birthDate": {
            "format": "date",
            "nullable": true,
            "type": "string"
          },
          "color": {
            "description": "This property was added in v1.126.0",
            "type": "string"
          },
          "faces": {
            "items": {
              "$ref": "#/components/schemas/AssetFaceWithoutPersonResponseDto"
            },
            "type": "array"
          },
          "id": {
            "type": "string"
          },
          "isFavorite": {
            "description": "This property was added in v1.126.0",
            "type": "boolean"
          },
          "isHidden": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "thumbnailPath": {
            "type": "string"
          },
          "updatedAt": {
            "description": "This property was added in v1.107.0",
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "birthDate",
          "faces",
          "id",
          "isHidden",
          "name",
          "thumbnailPath"
        ],
        "type": "object"
      },
...
      "SourceType": {
        "enum": [
          "machine-learning",
          "exif"
        ],
        "type": "string"
      },
...
      "SyncStreamDto": {
        "properties": {
          "types": {
            "items": {
              "enum": [
                "UserV1",
                "UserDeleteV1"
              ],
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "types"
        ],
        "type": "object"
      },
...
      "TagResponseDto": {
        "properties": {
          "color": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "parentId": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        },
        "required": [
          "createdAt",
          "id",
          "name",
          "updatedAt",
          "value"
        ],
        "type": "object"
      },
...
      "UserAvatarColor": {
        "enum": [
          "primary",
          "pink",
          "red",
          "yellow",
          "blue",
          "green",
          "purple",
          "orange",
          "gray",
          "amber"
        ],
        "type": "string"
      },
...
      "UserResponseDto": {
        "properties": {
          "avatarColor": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserAvatarColor"
              }
            ]
          },
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "profileChangedAt": {
            "format": "date-time",
            "type": "string"
          },
          "profileImagePath": {
            "type": "string"
          }
        },
        "required": [
          "avatarColor",
          "email",
          "id",
          "name",
          "profileChangedAt",
          "profileImagePath"
        ],
        "type": "object"
      },
...
    }
  }
}